separate chemistry helpers
